管理者チェックを追加

+++ src/main/java/actions/EmployeeAction.java
@@ -36,27 +36,26 @@ public class EmployeeAction extends ActionBase {
      */
     public void index() throws ServletException,IOException{
 
-        //指定されたページ数の一覧画面に表示するデータを取得
-        int page=getPage();
-        List<EmployeeView> employees=service.getPerPage(page);
-
-        //全ての従業員データの件数を取得
-        long employeeCount=service.countAll();
-
-        putRequestScope(AttributeConst.EMPLOYEES,employees);
-        putRequestScope(AttributeConst.EMP_COUNT,employeeCount);
-        putRequestScope(AttributeConst.PAGE,page);
-        putRequestScope(AttributeConst.MAX_ROW,JpaConst.ROW_PER_PAGE);
-
-        //セッションにフラッシュメッセージが設定されている場合はリクエストスコープに移し替え、セッションスコープを削除
-        String flush=getSessionScope(AttributeConst.FLUSH);
-        if(flush!=null) {
-            putRequestScope(AttributeConst.FLUSH,flush);
-            removeSessionScope(AttributeConst.FLUSH);
+        //管理者かどうかチェック
+        if (checkAdmin()) {
+            //指定されたページ数の一覧画面に表示するデータを取得
+            int page = getPage();
+            List<EmployeeView> employees = service.getPerPage(page);
+            //全ての従業員データの件数を取得
+            long employeeCount = service.countAll();
+            putRequestScope(AttributeConst.EMPLOYEES, employees);
+            putRequestScope(AttributeConst.EMP_COUNT, employeeCount);
+            putRequestScope(AttributeConst.PAGE, page);
+            putRequestScope(AttributeConst.MAX_ROW, JpaConst.ROW_PER_PAGE);
+            //セッションにフラッシュメッセージが設定されている場合はリクエストスコープに移し替え、セッションスコープを削除
+            String flush = getSessionScope(AttributeConst.FLUSH);
+            if (flush != null) {
+                putRequestScope(AttributeConst.FLUSH, flush);
+                removeSessionScope(AttributeConst.FLUSH);
+            }
+            //一覧画面を表示
+            forward(ForwardConst.FW_EMP_INDEX);
         }
-
-        //一覧画面を表示
-        forward(ForwardConst.FW_EMP_INDEX);
     }
 
     /**
@@ -65,11 +64,13 @@ public class EmployeeAction extends ActionBase {
      * @throws IOException
      */
     public void entryNew() throws ServletException,IOException{
-        putRequestScope(AttributeConst.TOKEN,getTokenId());//CSRF対策トークン
-        putRequestScope(AttributeConst.EMPLOYEE,new EmployeeView());//空の従業員インスタンス
-
-        //新規登録画面を表示
-        forward(ForwardConst.FW_EMP_NEW);
+        //管理者かどうかチェック
+        if (checkAdmin()) {
+            putRequestScope(AttributeConst.TOKEN, getTokenId());//CSRF対策トークン
+            putRequestScope(AttributeConst.EMPLOYEE, new EmployeeView());//空の従業員インスタンス
+            //新規登録画面を表示
+            forward(ForwardConst.FW_EMP_NEW);
+        }
     }
 
     /**
@@ -78,40 +79,42 @@ public class EmployeeAction extends ActionBase {
      * @throws IOException
      */
     public void create() throws ServletException,IOException{
-        //CSRF対策　tokenのチェック
-        if(checkToken()) {
-            //パラメータの値をもとに従業員情報のインスタンスを作成する
-            EmployeeView ev=new EmployeeView(
-                    null,
-                    getRequestParam(AttributeConst.EMP_CODE),
-                    getRequestParam(AttributeConst.EMP_NAME),
-                    getRequestParam(AttributeConst.EMP_PASS),
-                    toNumber(getRequestParam(AttributeConst.EMP_ADMIN_FLG)),
-                    null,
-                    null,
-                    AttributeConst.DEL_FLAG_FALSE.getIntegerValue()
-                    );
+        //管理者かどうかチェック
+        if (checkAdmin()) {
+            //CSRF対策　tokenのチェック
+            if (checkToken()) {
+                //パラメータの値をもとに従業員情報のインスタンスを作成する
+                EmployeeView ev = new EmployeeView(
+                        null,
+                        getRequestParam(AttributeConst.EMP_CODE),
+                        getRequestParam(AttributeConst.EMP_NAME),
+                        getRequestParam(AttributeConst.EMP_PASS),
+                        toNumber(getRequestParam(AttributeConst.EMP_ADMIN_FLG)),
+                        null,
+                        null,
+                        AttributeConst.DEL_FLAG_FALSE.getIntegerValue());
 
-            //アプリケーションスコープからpepper文字列を取得
-            String pepper=getContextScope(PropertyConst.PEPPER);
+                //アプリケーションスコープからpepper文字列を取得
+                String pepper = getContextScope(PropertyConst.PEPPER);
 
-            //従業員情報登録
-            List<String> errors=service.creat(ev, pepper);
+                //従業員情報登録
+                List<String> errors = service.creat(ev, pepper);
 
-            if(errors.size()>0) {
-                //エラーがあった場合
-                putRequestScope(AttributeConst.TOKEN,getTokenId());//CSRF対策トークン
-                putRequestScope(AttributeConst.EMPLOYEE,ev);//入力された従業員情報
-                putRequestScope(AttributeConst.ERR,errors);//エラーのリスト
+                if (errors.size() > 0) {
+                    //エラーがあった場合
+                    putRequestScope(AttributeConst.TOKEN, getTokenId());//CSRF対策トークン
+                    putRequestScope(AttributeConst.EMPLOYEE, ev);//入力された従業員情報
+                    putRequestScope(AttributeConst.ERR, errors);//エラーのリスト
 
-                //新規登録画面を再表示
-                forward(ForwardConst.FW_EMP_NEW);
-            }else {
-                //登録中にエラーがなかった場合
-                //セッションに登録完了のフラッシュメッセージを設定
-                putSessionScope(AttributeConst.FLUSH,MessageConst.I_REGISTERED.getMessage());
-                //一覧画面にリダイレクト
-                redirect(ForwardConst.ACT_EMP,ForwardConst.CMD_INDEX);
+                    //新規登録画面を再表示
+                    forward(ForwardConst.FW_EMP_NEW);
+                } else {
+                    //登録中にエラーがなかった場合
+                    //セッションに登録完了のフラッシュメッセージを設定
+                    putSessionScope(AttributeConst.FLUSH, MessageConst.I_REGISTERED.getMessage());
+                    //一覧画面にリダイレクト
+                    redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
+                }
             }
         }
     }
@@ -122,19 +125,19 @@ public class EmployeeAction extends ActionBase {
      * @throws IOException
      */
     public void show() throws ServletException,IOException{
-        //idを条件に従業員データを取得する
-        EmployeeView ev=service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
-
-        if(ev==null||ev.getDeleteFlag()==AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
-            //データが取得できなかった、または論理削除されいた場合はエラー画面を表示
-            forward(ForwardConst.FW_ERR_UNKNOWN);
-            return;
+        //管理者かどうかチェック
+        if (checkAdmin()) {
+            //idを条件に従業員データを取得する
+            EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
+            if (ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
+                //データが取得できなかった、または論理削除されいた場合はエラー画面を表示
+                forward(ForwardConst.FW_ERR_UNKNOWN);
+                return;
+            }
+            putRequestScope(AttributeConst.EMPLOYEE, ev);//取得した従業員情報
+            //詳細画面を表示
+            forward(ForwardConst.FW_EMP_SHOW);
         }
-
-        putRequestScope(AttributeConst.EMPLOYEE,ev);//取得した従業員情報
-
-        //詳細画面を表示
-        forward(ForwardConst.FW_EMP_SHOW);
     }
 
     /**
@@ -143,20 +146,20 @@ public class EmployeeAction extends ActionBase {
      * @throws IOException
      */
     public void edit() throws ServletException,IOException{
-        //idを条件に従業員データを取得する
-        EmployeeView ev=service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
-
-        if(ev==null||ev.getDeleteFlag()==AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
-            //データが取得できなかった、または論理削除されていた場合はエラー画面を表示
-            forward(ForwardConst.FW_ERR_UNKNOWN);
-            return;
+        //管理者かどうかチェック
+        if (checkAdmin()) {
+            //idを条件に従業員データを取得する
+            EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
+            if (ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
+                //データが取得できなかった、または論理削除されていた場合はエラー画面を表示
+                forward(ForwardConst.FW_ERR_UNKNOWN);
+                return;
+            }
+            putRequestScope(AttributeConst.TOKEN, getTokenId());//CSRF対策用トークン
+            putRequestScope(AttributeConst.EMPLOYEE, ev);//取得した従業員情報
+            //編集画面を表示する
+            forward(ForwardConst.FW_EMP_EDIT);
         }
-
-        putRequestScope(AttributeConst.TOKEN,getTokenId());//CSRF対策用トークン
-        putRequestScope(AttributeConst.EMPLOYEE,ev);//取得した従業員情報
-
-        //編集画面を表示する
-        forward(ForwardConst.FW_EMP_EDIT);
     }
 
     /**
@@ -165,41 +168,43 @@ public class EmployeeAction extends ActionBase {
      * @throws IOException
      */
     public void update() throws ServletException,IOException{
-        //CSRF対策トークンのチェック
-        if(checkToken()) {
-            //パラメータの値をもとに従業員情報のインスタスを作成する
-            EmployeeView ev=new EmployeeView(
-                    toNumber(getRequestParam(AttributeConst.EMP_ID)),
-                    getRequestParam(AttributeConst.EMP_CODE),
-                    getRequestParam(AttributeConst.EMP_NAME),
-                    getRequestParam(AttributeConst.EMP_PASS),
-                    toNumber(getRequestParam(AttributeConst.EMP_ADMIN_FLG)),
-                    null,
-                    null,
-                    AttributeConst.DEL_FLAG_FALSE.getIntegerValue()
-                    );
+        //管理者かどうかチェック
+        if (checkAdmin()) {
+            //CSRF対策トークンのチェック
+            if (checkToken()) {
+                //パラメータの値をもとに従業員情報のインスタスを作成する
+                EmployeeView ev = new EmployeeView(
+                        toNumber(getRequestParam(AttributeConst.EMP_ID)),
+                        getRequestParam(AttributeConst.EMP_CODE),
+                        getRequestParam(AttributeConst.EMP_NAME),
+                        getRequestParam(AttributeConst.EMP_PASS),
+                        toNumber(getRequestParam(AttributeConst.EMP_ADMIN_FLG)),
+                        null,
+                        null,
+                        AttributeConst.DEL_FLAG_FALSE.getIntegerValue());
 
-            //アプリケーションスコープからpepper文字列を取得
-            String pepper=getContextScope(PropertyConst.PEPPER);
+                //アプリケーションスコープからpepper文字列を取得
+                String pepper = getContextScope(PropertyConst.PEPPER);
 
-            //従業員情報更新
-            List<String> errors=service.update(ev, pepper);
+                //従業員情報更新
+                List<String> errors = service.update(ev, pepper);
 
-            if(errors.size()>0) {
-                //更新中にエラーが発生した場合
+                if (errors.size() > 0) {
+                    //更新中にエラーが発生した場合
 
-                putRequestScope(AttributeConst.TOKEN,getTokenId());//CSRF対策トークン
-                putRequestScope(AttributeConst.EMPLOYEE,ev);//入力された従業員情報
-                putRequestScope(AttributeConst.ERR,errors);//エラーのリスト
+                    putRequestScope(AttributeConst.TOKEN, getTokenId());//CSRF対策トークン
+                    putRequestScope(AttributeConst.EMPLOYEE, ev);//入力された従業員情報
+                    putRequestScope(AttributeConst.ERR, errors);//エラーのリスト
 
-                //編集画面を再表示
-                forward(ForwardConst.FW_EMP_EDIT);
-            }else {
-                //更新中にエラーがなかったら
-                //セッションに更新完了のフラッシュメッセージを設定
-                putSessionScope(AttributeConst.FLUSH,MessageConst.I_UPDATED.getMessage());
-                //一覧画面にリダイレクト
-                redirect(ForwardConst.ACT_EMP,ForwardConst.CMD_INDEX);
+                    //編集画面を再表示
+                    forward(ForwardConst.FW_EMP_EDIT);
+                } else {
+                    //更新中にエラーがなかったら
+                    //セッションに更新完了のフラッシュメッセージを設定
+                    putSessionScope(AttributeConst.FLUSH, MessageConst.I_UPDATED.getMessage());
+                    //一覧画面にリダイレクト
+                    redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
+                }
             }
         }
     }
@@ -210,14 +215,36 @@ public class EmployeeAction extends ActionBase {
      * @throws IOException
      */
     public void destroy() throws ServletException,IOException{
-        //CSRF対策トークンのチェック
-        if(checkToken()) {
-            //idを条件に従業員データを論理削除する
-            service.destroy(toNumber(getRequestParam(AttributeConst.EMP_ID)));
-            //セッションに削除完了のフラッシュメッセージを設定
-            putSessionScope(AttributeConst.FLUSH,MessageConst.I_DELETED.getMessage());
-            //一覧画面にリダイレクト
-            redirect(ForwardConst.ACT_EMP,ForwardConst.CMD_INDEX);
+        //管理者かどうかチェック
+        if (checkAdmin()) {
+            //CSRF対策トークンのチェック
+            if (checkToken()) {
+                //idを条件に従業員データを論理削除する
+                service.destroy(toNumber(getRequestParam(AttributeConst.EMP_ID)));
+                //セッションに削除完了のフラッシュメッセージを設定
+                putSessionScope(AttributeConst.FLUSH, MessageConst.I_DELETED.getMessage());
+                //一覧画面にリダイレクト
+                redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
+            }
+        }
+    }
+
+    /**
+     * ログイン中の従業員が管理者かどうかチェックし、管理者でなければエラー画面を表示
+     * true: 管理者 false: 管理者ではない
+     * @throws ServletException
+     * @throws IOException
+     */
+    public boolean checkAdmin() throws ServletException,IOException{
+        //セッションからログイン中の従業員情報を取得
+        EmployeeView ev=(EmployeeView)getSessionScope(AttributeConst.LOGIN_EMP);
+
+        //管理者でなければエラー画面を表示
+        if(ev.getAdminFlag()!=AttributeConst.ROLE_ADMIN.getIntegerValue()) {
+            forward(ForwardConst.FW_ERR_UNKNOWN);
+            return false;
+        }else {
+            return true;
         }
     }
 }
